Implementado melhorias em ordem de servico e cards de producao

+++ src/main/java/br/com/hitbox/infra/entity/ItemProductEntity.java
@@ -56,5 +56,5 @@ public class ItemProductEntity {
     @Column(nullable = false, precision = 19, scale = 2)
     private BigDecimal totalSalePrice;
 
-    private Long estimatedMinutes;
+    private BigDecimal estimatedMinutes;
 }
