+++ src/main/java/br/com/hitbox/interfaces/dto/request/order/ServiceOrderItemProductRequest.java
@@ -19,5 +19,5 @@ public class ServiceOrderItemProductRequest {
 
     private BigDecimal salePriceUnit;
 
-    private Long estimatedMinutes;
+    private BigDecimal estimatedMinutes;
 }
