+++ src/main/java/br/com/hitbox/interfaces/dto/response/order/ServiceOrderItemProductResponse.java
@@ -29,5 +29,5 @@ public class ServiceOrderItemProductResponse {
 
     private BigDecimal totalSalePrice;
 
-    private Long estimatedMinutes;
+    private BigDecimal estimatedMinutes;
 }
